fix: Reselection of entity
