+++ src/ai/backend/manager/actions/action.py
@@ -3,8 +3,6 @@ from dataclasses import dataclass
 from datetime import datetime
 from typing import Generic, Mapping, Optional, TypeVar
 
-from ai.backend.common.types import DispatchResult
-
 
 @dataclass
 class BaseAction(ABC):
@@ -38,21 +36,6 @@ class BaseBatchAction(ABC):
 class BaseActionResult(ABC):
     @abstractmethod
     def entity_id(self) -> Optional[str]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def status(self) -> str:
-        raise NotImplementedError
-
-    @abstractmethod
-    def description(self) -> Optional[str]:
-        raise NotImplementedError
-
-    def to_bgtask_result(self) -> DispatchResult:
-        """
-        Implement this method to convert the Result to DispatchResult.
-        You should implment this method for running the action as bgtask.
-        """
         raise NotImplementedError
