+++ src/ai/backend/manager/services/group/BUILD
