+++ src/ai/backend/manager/services/resource_preset/BUILD
