+++ src/ai/backend/manager/services/resource_preset/actions/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="src")
