+++ src/ai/backend/manager/services/resource_preset/base.py
@@ -1,0 +1,18 @@
+from dataclasses import dataclass
+from typing import override
+
+from ai.backend.manager.actions.action import BaseAction, BaseBatchAction
+
+
+@dataclass
+class ResourcePresetAction(BaseAction):
+    @override
+    def entity_type(self):
+        return "resource_preset"
+
+
+@dataclass
+class ResourcePresetBatchAction(BaseBatchAction):
+    @override
+    def entity_type(self):
+        return "resource_preset"
