+++ src/ai/backend/manager/services/user/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="src")
